+++ CHANGELOG.md
@@ -5,8 +5,10 @@
   - Section 9: Error Handling & Regular Expressions
     - Error Handling with Try...Catch
     - Regular Expressions [Part 1] - Evaluation Functions
+    - Regular Expressions [Part 2] - Metacharacter Symbols
 - An example with error catching was added.
 - An example with evaluation functions was added.
+- An example with metacharacter symbols was added.
 
 ## September 21, 2020
 - Modern JavaScript From The Beginning (Udemy)